+++ src/components/Map/LocationTracker.tsx
@@ -12,14 +12,22 @@ type LocationTrackerProps = {
   setUserLocation: Dispatch<SetStateAction<LatLngExpression>>;
   locationPref: boolean | null;
   setLocationPref: Dispatch<SetStateAction<boolean | null>>;
+  inAddPointMode: boolean;
+  setAddpointCords: Dispatch<SetStateAction<LatLngTuple | null>>
 };
 
-function LocationTracker({ setUserLocation, setLocationPref, locationPref }: LocationTrackerProps) {
+function LocationTracker({ setUserLocation, setLocationPref, locationPref, inAddPointMode, setAddpointCords }: LocationTrackerProps) {
   const [userPos, setUserPos] = useState<LatLngTuple | undefined>(undefined);
   const trackingWatcherStarted = useRef(false);
   const initialTrack = useRef(true);
 
   const map = useMapEvents({
+    click(e) {
+      if (!inAddPointMode) {
+        return;
+      }
+      setAddpointCords([e.latlng.lat, e.latlng.lng]);
+    },
     locationfound(e: LocationEvent) {
       if (userPos !== undefined) {
         const latDist = e.latlng.lat - Number(userPos[0]);
@@ -56,7 +64,9 @@ function LocationTracker({ setUserLocation, setLocationPref, locationPref }: Loc
       trackingWatcher();
       trackingWatcherStarted.current = true;
       // start interval
-      setInterval(trackingWatcher, 20000);
+      const trackingInterval = setInterval(trackingWatcher, 20000);
+
+      return () => { clearInterval(trackingInterval) }
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);