+++ src/components/Map/MapPopup.tsx
@@ -17,9 +17,21 @@ function MapPopup({ ...props }: MapPopupProps) {
 
   const url: string = getDeviceMapProvider() + props.pin.latitude + ',' + props.pin.longitude;
 
+  function getBgImg(): string {
+    return (props.pin.image[0]?.cloud_pic === undefined || props?.pin?.image[0].cloud_pic === "") ? '/assets/mappoint-placeholder-img.jpg' : props?.pin?.image[0].cloud_pic;
+  }
+
+  function isImagePlaceholder(): boolean {
+    return (props.pin.image[0]?.cloud_pic === undefined || props?.pin?.image[0].cloud_pic === "");
+  }
+
   return (
     <Popup>
-      <div id="map-popup-container">
+      <div
+        style={{ backgroundImage: `url('${getBgImg()}')` }}
+        className={`map-popup-image w-full h-28 bg-no-repeat bg-cover bg-dark-seaweed ${isImagePlaceholder() ? 'bg-bottom' : 'bg-center'}`}
+      ></div>
+      <div id="map-popup-container" className="my-3.5 mr-6 ml-5">
         <h3 className="font-bold text-default-font text-md">{props.pin.title}</h3>
         <h4 className="mt-1 text-bright-seaweed">{props.pin.category}</h4>
         <p className="font-light text-default-font">