+++ src/types/Pins/RatingProps.d.ts
@@ -1,0 +1,6 @@
+type RatingProps = {
+  id: number;
+  rating: number;
+  mappoint: string;
+  creator: string;
+};