+++ src/util/capitalizeFirstLetter.ts
@@ -1,0 +1,8 @@
+function capitalizeFirstLetter(text: string) {
+  if (text.length === 0) {
+    return '';
+  }
+  return text.charAt(0).toUpperCase() + text.slice(1);
+}
+
+export { capitalizeFirstLetter }