+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - pages are connected correctly
 
 ### v1.0.0
-- profil page
+- profile page
 - updated map page + features
 - navbar
 - auth is connected to backend
@@ -26,7 +26,7 @@
 - remove unused libraries
 
 ### v1.2.0
-- Add change profil data
+- Add change profile data
 - Add live tracking
 - Add password forgot
 - Add User email verification
@@ -34,3 +34,16 @@
 
 ### v1.2.1
 - Fix iphone map scrolling
+
+### v2.0.0
+- Add favorite page
+- Add Image upload for user profile
+- Allow to filter by favorites on map
+- Fix an inconsisten data bug caused by getServerSideProps on map & profile page
+- Add bookmark (favorite) function on mappoints
+- Adding & deleting comments
+- Rating a mappoint
+- Show a profile picture on each comment
+- User can now add mappoints
+- map pin popup preview now shows an image
+- mappoints now have a slider if several images are present for that spot